add(PMI2): add PMI2 support
LCI is built on PMIv1 wire protocol but PMIv2 API, so the underlying system should provide PMI2 shared library (libpmi2.so or libpmi.so on some Cray machine) for LCI.

Users need to select PMI/PMI2 when compiling LCI with the cmake option `LCI_USE_PMI2`. If LCI cannot find libpmi2.so/libpmi.so, it will switch back to PMIv1.

LCI_PM_Barrier() is broken with PMI2