proto: register/dereg mem before/after RMA send
ibv registered send data in lc_server_putl/putls/rma_rtr,
but never deregistered the send data; with sufficient sends,
this could oveflow the registration cache.
Register/deregister instead as part of the rendezvous protocol;
this is necessary for RDMA networks anyway,
and is a no-op on others (e.g. psm2).