core: add send queue credits, better error checking
The primary mechanism for ensuring that send queues didn't overflow was
enforcing packet pool size == send queue size. Increasing the size of
the packet pool (to support more concurrent recieves) resulted in
overflows of the send queue. This patch adds explicit send queue
credits, currently supported in the IBVerbs backend, that ensures we
don't start a send if it will overflow the queue - instead, a RETRY
error is returned to the user.

This patch also adds better failure reporting to the user and, with
server debug mode enabled, can cause a hard error if problems occur when
they cannot be easily returned as a RETRY error.