Refactor performance counters: "on-the-fly" and "on-the-fly-lw" modes.
Rename the previous "on-the-fly" mode into "on-the-fly-lw": in this mode,
a monitor thread will tell application threads "it's time to dump the performance counters".
It is the application threads' job to dump the performance counters.
It is called "lightweight" because no performance counters will be dumped if they are not changed.

In the new "on-the-fly" mode, a monitor thread will directly dump
the performance counters at fixed intervals. It is useful for debugging deadlocks.